Update uid type

+++ app/src/main/java/onlymash/flexbooru/model/PostMoe.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "posts_moebooru", indices = [(Index(value = ["host", "keyword", "id"], unique = true))])
 data class PostMoe(
     @PrimaryKey(autoGenerate = true)
-    val uid: Int?,
+    val uid: Long?,
     var host: String?,
     var keyword: String?,
     val id: Int,
